wall_follow_multi.cpp: change variable val from type int8_t to int16_t.
This variable is for storing the return status of path_escape_trapped(),
which is updated in the previous commit:

	commit f77fa929a40166b7d40177b198c408ca160c6843
	Author: Patrick Chau <patrickchau@china-ilife.com>
	Date:   Sat Nov 19 11:11:01 2016 +0800

	    Change return type of the following functions to int16_t instead of
	    int8_t, since the map size might be overflow.

	    path_planning.cpp: path_escape_trapped()
	    shortest_path.cpp: path_find_shortest_path(), path_move_to_cell(),
	    path_move_to_unclean_area()

+++ src/wall_follow_multi.cpp
@@ -108,7 +108,7 @@ void WFM_move_back(uint16_t dist)
 void *WFM_check_trapped(void *data)
 {
 	float		pos_x, pos_y;
-	int8_t		val;
+	int16_t		val;
 	uint32_t	left_speed, right_speed;
 	static int16_t	current_x = 0, current_y = 0;
